renamed initPermissionConfig.js to initPermissionConfigFile.js
