+++ server/helpers/initPermissionConfigFile.js
